@@ -1,16 +1,15 @@
-package java.ru.javawebinar.topjava.repository.inmemory;
+package ru.javawebinar.topjava.repository.inmemory;
 
 import org.springframework.stereotype.Repository;
-import java.ru.javawebinar.topjava.UserTestData;
+import ru.javawebinar.topjava.UserTestData;
 import ru.javawebinar.topjava.model.User;
 import ru.javawebinar.topjava.repository.UserRepository;
 
 import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.Collectors;
 
-import static java.ru.javawebinar.topjava.UserTestData.*;
+import static ru.javawebinar.topjava.UserTestData.*;
 
 
 @Repository
@@ -28,7 +27,7 @@ public class InMemoryUserRepository extends InMemoryBaseRepository<User> impleme
     public List<User> getAll() {
         return getCollection().stream()
                 .sorted(Comparator.comparing(User::getName).thenComparing(User::getEmail))
-                .collect(Collectors.toList());
+                .toList();
     }
 
     @Override
