@@ -1,4 +1,4 @@
-package java.ru.javawebinar.topjava;
+package ru.javawebinar.topjava;
 
 import ru.javawebinar.topjava.model.Role;
 import ru.javawebinar.topjava.model.User;
