@@ -1,4 +1,4 @@
-package java.ru.javawebinar.topjava.repository.inmemory;
+package ru.javawebinar.topjava.repository.inmemory;
 
 import ru.javawebinar.topjava.model.AbstractBaseEntity;
 
